Improve responsibility when canceling downloading file.
If downloading speed is very very slow, user should wait quite long time
  until finishing one-read-operation (until download buffer is filled)
   becase of 'synchonzied' lock.
This may takes longer than expected depending on network state.
So, in spite of risk of stability, 'lock' for downloading file is removed,
  and just 'boolean flag' is used for it.
I'm not sure about stability due to race-condition but it seems ok for me.
I need to keep eyes on stablity in this case...