Fix bug - Race condition
Sometimes, SQLite issues exception something like this, "There is no such table ...".
Because, at the end of update process, current item table is dropped and new item table is renamed.
But, db request to this table between drop and rename, is fails because there is not table!
So, DB operations related with this issue are synchronized.