Initialize routine move to FeederApp.java / License term is added.
Why initialization routine is moved to FeederApp again?
May error reports says about app is not initialized.
So, in some use cases, I think Android FW start's specific activity
  directly even if process is finished.
(Saving app state and restore it something like this.)
So, let's move it to App. And I need to monitor it.