Fix BIG MISTAKE of algorithm to decide action type by analyze parsed feed item.
There was BIG MISTAKE when adding special action for Youtube.
After the Youtube action change, every newly added podCast channel uses 'OPEN' action (NOT 'DOWNLOAD')!!
So, this mistake is fixed and additionaly, now action is decided whenever channel update is executed.
Why?
Even if it's very rare, whole item type of channel may be changed later.
Or in case of some channels, there may be NO items at first and then item is added later.
I'm not sure that this - updating action at every channel update - is better than before or not, and it doens't have any side-effect or not... but let's try.