Notification algorithm - checking & intent - is changed
enQ/deQ based notification has bug when task is cancelled before it really starting.
If it cancelled before starting, notification misunderstanding that one running task is cancelled even if it isn't even started.
This leads to unexpected state of NotiManager.
So, now NotiManager check bg task when it is started - before entering readyQ/runQ -  and it is endded.

And, There is some serious issues regaring CONTENT_ACTION.
I can't find a way to tell wether Feeder is top application or not.
So, CONTENT_ACTION is disable - commented out